# Simplify and optimize some GraphNode find functions