Follow-up of #4603

This PR simplify `GraphNode.find` and `GraphNode.findOne`.
It also avoids creating and merging lots of arrays in `GraphNode.find` which means less garbage.
In addition, `GraphNode.findByName` now simply calls `GraphNode.findOne` so the traversing logic isn't duplicated anymore.

I confirm I have read the [contributing guidelines](https://github.com/playcanvas/engine/blob/master/.github/CONTRIBUTING.md) and signed the [Contributor License Agreement](https://docs.google.com/a/playcanvas.com/forms/d/1Ih69zQfJG-QDLIEpHr6CsaAs6fPORNOVnMv5nuo0cjk/viewform).