@Maksims 2023-11-25
Do you have a benchmarks with various hierarchy cases?