@NewboO 2023-11-27
> Do you have a benchmarks with various hierarchy cases?

I ran somes on one of my projects which has a hierarchy of ~800 nodes.
Benchmarked from root and from a deeper node (only 80 descendants), finding by name both existing (100 occurrences to return from root, 5 from the deeper node) and non-existing:

```
GraphNode.find, existing name

                       |  old  |  new
--------------------------------------
800+ nodes (100k runs) |  5.3s |  2.5s
--------------------------------------
80 nodes (1M runs)     | 0.98s | 0.75s


GraphNode.find, non-existing name

                       |  old  |  new
--------------------------------------
800+ nodes (100k runs) |  4.6s |  2.5s
--------------------------------------
80 nodes (1M runs)     | 0.63s | 0.70s


GraphNode.findOne, existing name relatively deep

                       |   old  |   new
----------------------------------------
800+ nodes (100k runs) |  0.29s |  0.17s
----------------------------------------
80 nodes (1M runs)     | 0.027s | 0.036s


GraphNode.findOne, non-existing name

                       |  old  |  new
--------------------------------------
800+ nodes (100k runs) |  4.5s |  2.7s
--------------------------------------
80 nodes (1M runs)     | 0.56s | 0.69s
```

So it may be sometimes a bit slower when only a few nodes need to be checked, but it gets way faster on bigger hierarchies.