@slimbuck 2023-11-28
This PR seems to me a wonderful example of how simplifying code can make it more readable, maintainable and run faster all at the same time. Great job @NewboO !!